feat(preprocess-executor): refactor preprocessing is handled by app
