+++ now/apps/base/app.py
@@ -217,3 +217,13 @@ class JinaNOWApp:
         :param app_config: contains all information needed to clean up the allocated resources
         """
         pass
+
+    def load_from_folder(self, path: str) -> DocumentArray:
+        """Loads files from folder into DocumentArray without any preprocessing or actual loading of documents."""
+        raise Exception(
+            f'modality {self.output_modality} not supported for data loading from folder'
+        )
+
+    def preprocess(self, da: DocumentArray, user_input: UserInput) -> DocumentArray:
+        """Loads and preprocesses every document such that it is ready for finetuning/indexing."""
+        return da
