+++ now/apps/image_to_text/app.py
@@ -3,13 +3,14 @@ from typing import Dict, List
 
 from docarray import DocumentArray
 from now_common import options
-from now_common.utils import setup_clip_music_apps
+from now_common.utils import preprocess_text, setup_clip_music_apps
 
 from now.apps.base.app import JinaNOWApp
 from now.constants import (
     CLIP_USES,
     IMAGE_MODEL_QUALITY_MAP,
     Apps,
+    DatasetTypes,
     Modalities,
     Qualities,
 )
@@ -92,3 +93,18 @@ class ImageToText(JinaNOWApp):
             indexer_uses='DocarrayIndexerV2',
             kubectl_path=kubectl_path,
         )
+
+    def load_from_folder(self, path: str) -> DocumentArray:
+        return DocumentArray.from_files(path + '/*.txt')
+
+    def preprocess(self, da: DocumentArray, user_input: UserInput) -> DocumentArray:
+        split_by_sentences = False
+        if (
+            user_input.is_custom_dataset
+            and user_input.custom_dataset_type == DatasetTypes.PATH
+            and user_input.dataset_path
+            and os.path.isdir(user_input.dataset_path)
+        ):
+            # for text loaded from folder can't assume it is split by sentences
+            split_by_sentences = True
+        return preprocess_text(da=da, split_by_sentences=split_by_sentences)
