+++ now/apps/text_to_text/app.py
@@ -2,7 +2,7 @@ import os
 from typing import Dict
 
 from docarray import DocumentArray
-from now_common.utils import setup_clip_music_apps
+from now_common.utils import preprocess_text, setup_clip_music_apps
 
 from now.apps.base.app import JinaNOWApp
 from now.constants import Apps, Modalities, Qualities
@@ -77,3 +77,18 @@ class TextToText(JinaNOWApp):
             indexer_uses='DocarrayIndexerV2',
             kubectl_path=kubectl_path,
         )
+
+    def load_from_folder(self, path: str) -> DocumentArray:
+        return DocumentArray.from_files(path + '/*.txt')
+
+    def preprocess(self, da: DocumentArray, user_input: UserInput) -> DocumentArray:
+        split_by_sentences = False
+        if (
+            user_input.is_custom_dataset
+            and user_input.custom_dataset_type == DatasetTypes.PATH
+            and user_input.dataset_path
+            and os.path.isdir(user_input.dataset_path)
+        ):
+            # for text loaded from folder can't assume it is split by sentences
+            split_by_sentences = True
+        return preprocess_text(da=da, split_by_sentences=split_by_sentences)
