+++ now/run_backend.py
@@ -25,7 +25,8 @@ def run(app_instance: JinaNOWApp, user_input: UserInput, kubectl_path: str):
     :param kubectl_path:
     :return:
     """
-    dataset = load_data(app_instance.output_modality, user_input)
+    dataset = load_data(app_instance, user_input)
+    dataset = app_instance.preprocess(da=dataset, user_input=user_input)
     env_dict = app_instance.setup(
         dataset=dataset, user_input=user_input, kubectl_path=kubectl_path
     )
